+++ src/main/java/de/maxhenkel/camera/entities/ImageEntity.java
@@ -2,7 +2,6 @@ package de.maxhenkel.camera.entities;
 
 import de.maxhenkel.camera.Main;
 import de.maxhenkel.camera.gui.ResizeFrameScreen;
-import de.maxhenkel.camera.items.ImageItem;
 import de.maxhenkel.camera.net.MessageResizeFrame;
 import net.minecraft.client.Minecraft;
 import net.minecraft.entity.Entity;
@@ -102,7 +101,7 @@ public class ImageEntity extends Entity {
         }
 
         if (stack.getItem().equals(Main.IMAGE)) {
-            UUID uuid = ImageItem.getUUID(stack);
+            UUID uuid = Main.IMAGE.getUUID(stack);
             if (uuid == null) {
                 return true;
             }