+++ src/main/java/de/maxhenkel/camera/gui/ResizeFrameScreen.java
@@ -1,6 +1,7 @@
 package de.maxhenkel.camera.gui;
 
 import com.mojang.blaze3d.systems.RenderSystem;
+import de.maxhenkel.camera.Config;
 import de.maxhenkel.camera.Main;
 import de.maxhenkel.camera.entities.ImageEntity;
 import de.maxhenkel.camera.net.MessageResizeFrame;
@@ -9,20 +10,26 @@ import net.minecraft.client.gui.screen.inventory.ContainerScreen;
 import net.minecraft.client.gui.widget.button.Button;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.math.AxisAlignedBB;
+import net.minecraft.util.text.TextFormatting;
 import net.minecraft.util.text.TranslationTextComponent;
 
 import java.util.UUID;
 
 public class ResizeFrameScreen extends ContainerScreen {
 
-    private static final ResourceLocation CAMERA_TEXTURE = new ResourceLocation(Main.MODID, "textures/gui/camera.png");
-    private static final int FONT_COLOR = 4210752;
+    private static final ResourceLocation CAMERA_TEXTURE = new ResourceLocation(Main.MODID, "textures/gui/resize_frame.png");
+    private static final int PADDING = 10;
+    private static final int BUTTON_HEIGHT = 20;
+    private static final int BUTTON_WIDTH = 50;
 
     private UUID uuid;
+    private float visibility;
+    private Button visibilityButton;
 
     public ResizeFrameScreen(UUID uuid) {
         super(new DummyContainer(), null, new TranslationTextComponent("gui.frame.resize"));
         this.uuid = uuid;
+        visibility = Config.CLIENT.RESIZE_GUI_OPACITY.get().floatValue();
         xSize = 248;
         ySize = 109;
     }
@@ -33,24 +40,31 @@ public class ResizeFrameScreen extends ContainerScreen {
 
         buttons.clear();
         int left = (width - xSize) / 2;
-        int padding = 10;
-        int buttonWidth = 50;
-        int buttonHeight = 20;
-        addButton(new Button(left + padding, height / 2 - buttonHeight / 2, buttonWidth, buttonHeight, new TranslationTextComponent("button.frame.left").getFormattedText(), (button) -> {
+        addButton(new Button(left + PADDING, height / 2 - BUTTON_HEIGHT / 2, BUTTON_WIDTH, BUTTON_HEIGHT, "", (button) -> {
             sendMoveImage(MessageResizeFrame.Direction.LEFT);
         }));
 
-        addButton(new Button(left + xSize - buttonWidth - padding, height / 2 - buttonHeight / 2, buttonWidth, buttonHeight, new TranslationTextComponent("button.frame.right").getFormattedText(), (button) -> {
+        addButton(new Button(left + xSize - BUTTON_WIDTH - PADDING, height / 2 - BUTTON_HEIGHT / 2, BUTTON_WIDTH, BUTTON_HEIGHT, "", (button) -> {
             sendMoveImage(MessageResizeFrame.Direction.RIGHT);
         }));
 
-        addButton(new Button(width / 2 - buttonWidth / 2, guiTop + padding, buttonWidth, buttonHeight, new TranslationTextComponent("button.frame.up").getFormattedText(), (button) -> {
+        addButton(new Button(width / 2 - BUTTON_WIDTH / 2, guiTop + PADDING, BUTTON_WIDTH, BUTTON_HEIGHT, "", (button) -> {
             sendMoveImage(MessageResizeFrame.Direction.UP);
         }));
 
-        addButton(new Button(width / 2 - buttonWidth / 2, guiTop + ySize - padding - buttonHeight, buttonWidth, buttonHeight, new TranslationTextComponent("button.frame.down").getFormattedText(), (button) -> {
+        addButton(new Button(width / 2 - BUTTON_WIDTH / 2, guiTop + ySize - PADDING - BUTTON_HEIGHT, BUTTON_WIDTH, BUTTON_HEIGHT, "", (button) -> {
             sendMoveImage(MessageResizeFrame.Direction.DOWN);
         }));
+
+        visibilityButton = new Button(left + xSize - 20 - PADDING, guiTop + PADDING, 20, 20, new TranslationTextComponent("tooltip.visibility_short").getFormattedText(), (button) -> {
+            visibility -= 0.25;
+            if (visibility < 0F) {
+                visibility = 1F;
+            }
+            Config.CLIENT.RESIZE_GUI_OPACITY.set((double) visibility);
+            Config.CLIENT.RESIZE_GUI_OPACITY.save();
+        });
+        addButton(visibilityButton);
     }
 
     private void sendMoveImage(MessageResizeFrame.Direction direction) {
@@ -62,15 +76,37 @@ public class ResizeFrameScreen extends ContainerScreen {
         super.drawGuiContainerForegroundLayer(x, y);
 
         String title = new TranslationTextComponent("gui.frame.resize").getFormattedText();
-
         int titleWidth = font.getStringWidth(title);
-        font.drawString(title, xSize / 2 - titleWidth / 2, ySize / 2 - font.FONT_HEIGHT / 2, FONT_COLOR);
+        font.drawString(title, xSize / 2 - titleWidth / 2, ySize / 2 - font.FONT_HEIGHT - 1, TextFormatting.DARK_GRAY.getColor());
+
+        String description = new TranslationTextComponent("gui.frame.resize_description").getFormattedText();
+        int descriptionWidth = font.getStringWidth(description);
+        font.drawString(description, xSize / 2 - descriptionWidth / 2, ySize / 2 + 1, TextFormatting.GRAY.getColor());
+
+        minecraft.getTextureManager().bindTexture(CAMERA_TEXTURE);
+        if (Screen.hasShiftDown()) {
+            blit(xSize / 2 - 8, PADDING + 2, 16, 109, 16, 16);
+            blit(xSize / 2 - 8, ySize - PADDING - BUTTON_HEIGHT + 2, 0, 109, 16, 16);
+            blit(PADDING + BUTTON_WIDTH / 2 - 8, ySize / 2 - BUTTON_HEIGHT / 2 + 3, 0, 125, 16, 16);
+            blit(xSize - PADDING - BUTTON_WIDTH / 2 - 8, ySize / 2 - BUTTON_HEIGHT / 2 + 3, 16, 125, 16, 16);
+        } else {
+            blit(xSize / 2 - 8, PADDING + 2, 0, 109, 16, 16);
+            blit(xSize / 2 - 8, ySize - PADDING - BUTTON_HEIGHT + 2, 16, 109, 16, 16);
+            blit(PADDING + BUTTON_WIDTH / 2 - 8, ySize / 2 - BUTTON_HEIGHT / 2 + 3, 16, 125, 16, 16);
+            blit(xSize - PADDING - BUTTON_WIDTH / 2 - 8, ySize / 2 - BUTTON_HEIGHT / 2 + 3, 0, 125, 16, 16);
+        }
+
+        if (visibilityButton.isHovered()) {
+            renderTooltip(new TranslationTextComponent("tooltip.visibility").getFormattedText(), x - guiLeft, y - guiTop);
+        }
     }
 
     @Override
     protected void drawGuiContainerBackgroundLayer(float partialTicks, int mouseX, int mouseY) {
-        renderBackground();
-        RenderSystem.color4f(1.0F, 1.0F, 1.0F, 1.0F);
+        if (visibility >= 1F) {
+            renderBackground();
+        }
+        RenderSystem.color4f(1.0F, 1.0F, 1.0F, visibility);
         minecraft.getTextureManager().bindTexture(CAMERA_TEXTURE);
         blit(guiLeft, guiTop, 0, 0, xSize, ySize);
     }