+++ src/main/java/de/maxhenkel/camera/integration/jei/JEIPlugin.java
@@ -2,7 +2,9 @@ package de.maxhenkel.camera.integration.jei;
 
 import de.maxhenkel.camera.Main;
 import de.maxhenkel.camera.gui.AlbumScreen;
+import de.maxhenkel.camera.gui.CameraScreen;
 import de.maxhenkel.camera.gui.ImageScreen;
+import de.maxhenkel.camera.gui.ResizeFrameScreen;
 import mezz.jei.api.IModPlugin;
 import mezz.jei.api.registration.IGuiHandlerRegistration;
 import net.minecraft.util.ResourceLocation;
@@ -17,7 +19,9 @@ public class JEIPlugin implements IModPlugin {
 
     @Override
     public void registerGuiHandlers(IGuiHandlerRegistration registration) {
-        registration.addGuiContainerHandler(ImageScreen.class, new ImageContainerHandler<>());
-        registration.addGuiContainerHandler(AlbumScreen.class, new ImageContainerHandler<>());
+        registration.addGuiContainerHandler(ImageScreen.class, new NoJEIContainerHandler<>());
+        registration.addGuiContainerHandler(AlbumScreen.class, new NoJEIContainerHandler<>());
+        registration.addGuiContainerHandler(ResizeFrameScreen.class, new NoJEIContainerHandler<>());
+        registration.addGuiContainerHandler(CameraScreen.class, new NoJEIContainerHandler<>());
     }
 }